+++ healthz/externals/health_client.go
@@ -1,0 +1,23 @@
+package externals
+
+import (
+	"net/http"
+	"time"
+)
+
+func GetHealthCheck(url string) error {
+	req, err := http.NewRequest("GET", url, nil)
+	if err != nil {
+		return err
+	}
+
+	client := &http.Client{Timeout: 1 * time.Second}
+	resp, err := client.Do(req)
+	if err != nil {
+		return err
+	}
+
+	resp.Body.Close()
+
+	return nil
+}
