+++ healthz/services/health.go
@@ -1,10 +1,10 @@
-package externals
+package services
 
 import (
 	"fmt"
-	"go-healthcheck/dto"
+	"go-healthcheck/healthz/dto"
+	"go-healthcheck/healthz/externals"
 	"math/big"
-	"net/http"
 	"time"
 )
 
@@ -17,7 +17,7 @@ func GetHealthSummary(healths []dto.Health) (*dto.Summary, error) {
 	fmt.Println(r.Binomial(1000, 10))
 
 	for _, health := range healths {
-		err := getHealthCheck(health.URL)
+		err := externals.GetHealthCheck(health.URL)
 		if err != nil {
 			summary.Failure++
 			continue
@@ -33,21 +33,4 @@ func GetHealthSummary(healths []dto.Health) (*dto.Summary, error) {
 	summary.TotalTime = elapsed.Nanoseconds()
 
 	return summary, nil
-}
-
-func getHealthCheck(url string) error {
-	req, err := http.NewRequest("GET", url, nil)
-	if err != nil {
-		return err
-	}
-
-	client := &http.Client{Timeout: 1 * time.Second}
-	resp, err := client.Do(req)
-	if err != nil {
-		return err
-	}
-
-	resp.Body.Close()
-
-	return nil
 }
